api/requirements.txt
argh==0.26.2
astroid==2.4.2
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
influxdb-client==1.9.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.1
paho-mqtt==1.5.0
pathtools==0.1.2
pycodestyle==2.6.0
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
Rx==3.1.1
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
watchdog==0.10.3
Werkzeug==1.0.1
wrapt==1.12.1
flask-cors==3.0.8
Flask-SocketIO==4.3.1
eventlet==0.26.1
Flask-MQTT==1.1.1